Update on Animation, scene viewer, guizmo, sprite component, asset browser and resourcemanage for icon texture
TAKE IN THE IMGUI.INI as there is new feature added

Animation:
able to animate within the display time given but cannot set display time for each individual frame. Just for player for now.
Player:
able to toggle idle sprite to running sprite for now only. Need to set some enum class for facing/direction.

Guizmo:
need to act in more feature in camera2d, as out guizmo not optimize
Must create a pick and drop class to able to select the entity in the scene
ImGui:
Able to set boolean on is animating or not and whether is a spritesheet a not
Have a scene viewer but have to tell the input to happen within the scene viewer instead of window.
For content browser able to drag but no drop in the scene must optimize it.