animation done
able to change the timing according to the display time indicate in the frame.
max frame is number of frame set by user.
TAKE NOTE the downside is that once the max frame is set, not able to alter it.
able to change the displaytime during runtime. and each individual sprite running in it own internaltime. no global time is needed.